refactor: share VoiceRow, Meter, StatPair, Toggle across onboarding and models
Extract the PR #91 VoiceRow component (with speed/quality meters,
cost badges, streaming indicators, download progress) into a shared
components/voice-row.tsx module. Move VoiceItem type, VOICE_META,
buildVoiceItems, and related constants into lib/models.ts.

The onboarding voice-model step now uses the same VoiceRow component
as the settings models page — unified local + cloud list with rich
metadata instead of the previous flat lists behind a Cloud/Local
toggle. The LLM cleanup step uses the shared Toggle component.
