+++ apps/electron/src/renderer/src/pages/onboarding.tsx
@@ -2,12 +2,11 @@ import { apiKeySchema } from "@freestyle/validations";
 import { zodResolver } from "@hookform/resolvers/zod";
 import markDark from "@renderer/assets/mark-dark.svg";
 import markLight from "@renderer/assets/mark-light.svg";
+import { Toggle, VoiceRow } from "@renderer/components/voice-row";
 import { getApiBase, getClient } from "@renderer/lib/api";
 import {
   type AvailableModel,
-  CLOUD_VOICE_PROVIDERS,
-  formatBytes,
-  formatSpeed,
+  buildVoiceItems,
   LLM_PROVIDERS,
   PROVIDER_DISPLAY_NAMES,
   type WhisperStatus,
@@ -18,15 +17,12 @@ import {
   Check,
   ChevronLeft,
   ChevronRight,
-  Download,
   ExternalLink,
   Eye,
   EyeOff,
-  HardDrive,
   Keyboard,
   Loader2,
   Mic,
-  RefreshCw,
   Shield,
   Sparkles,
 } from "lucide-react";
@@ -37,8 +33,6 @@ import { useNavigate } from "react-router";
 type Step = "welcome" | "permissions" | "voice-model" | "llm-cleanup";
 
 const STEPS: Step[] = ["welcome", "permissions", "voice-model", "llm-cleanup"];
-
-type ModelSource = "cloud" | "local";
 
 const IS_MAC =
   typeof navigator !== "undefined" && navigator.userAgent.includes("Mac");
@@ -52,11 +46,13 @@ export default function OnboardingPage(): React.JSX.Element {
   const [accessibilityStatus, setAccessibilityStatus] = useState(false);
 
   // Voice model state
-  const [modelSource, setModelSource] = useState<ModelSource>("cloud");
   const [available, setAvailable] = useState<AvailableModel[]>([]);
   const [selectedModel, setSelectedModel] = useState<AvailableModel | null>(
     null,
   );
+  const [selectedWhisperDefId, setSelectedWhisperDefId] = useState<
+    string | null
+  >(null);
   const apiKeyForm = useForm<{ provider: string; key: string }>({
     resolver: zodResolver(apiKeySchema),
     defaultValues: { provider: "", key: "" },
@@ -71,9 +67,6 @@ export default function OnboardingPage(): React.JSX.Element {
   const [whisperStatus, setWhisperStatus] = useState<WhisperStatus | null>(
     null,
   );
-  const [selectedWhisperModel, setSelectedWhisperModel] = useState<
-    string | null
-  >(null);
 
   // LLM cleanup state
   const [llmCleanup, setLlmCleanup] = useState(false);
@@ -178,10 +171,10 @@ export default function OnboardingPage(): React.JSX.Element {
     setTimeout(() => clearInterval(interval), 30000);
   }, []);
 
-  const selectModel = useCallback(
+  const selectCloudModel = useCallback(
     (model: AvailableModel) => {
       setSelectedModel(model);
-      setSelectedWhisperModel(null);
+      setSelectedWhisperDefId(null);
       if (!apiKeys.has(model.provider_id)) {
         setNeedsKey(true);
         apiKeyForm.reset({ provider: model.provider_id, key: "" });
@@ -192,8 +185,8 @@ export default function OnboardingPage(): React.JSX.Element {
     [apiKeys, apiKeyForm],
   );
 
-  const selectLocalWhisper = useCallback((modelId: string) => {
-    setSelectedWhisperModel(modelId);
+  const selectLocalModel = useCallback((defId: string, _name: string) => {
+    setSelectedWhisperDefId(defId);
     setSelectedModel(null);
     setNeedsKey(false);
   }, []);
@@ -255,9 +248,9 @@ export default function OnboardingPage(): React.JSX.Element {
             is_default: true,
           },
         });
-      } else if (selectedWhisperModel && whisperStatus) {
+      } else if (selectedWhisperDefId && whisperStatus) {
         const def = whisperStatus.modelDefinitions.find(
-          (d) => d.id === selectedWhisperModel,
+          (d) => d.id === selectedWhisperDefId,
         );
         if (def) {
           await client.api.models.configured.$post({
@@ -272,7 +265,7 @@ export default function OnboardingPage(): React.JSX.Element {
           fetch(`${getApiBase()}/api/whisper/server/start`, {
             method: "POST",
             headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ modelId: selectedWhisperModel }),
+            body: JSON.stringify({ modelId: selectedWhisperDefId }),
           }).catch(() => {});
         }
       }
@@ -285,7 +278,7 @@ export default function OnboardingPage(): React.JSX.Element {
     }
   }, [
     selectedModel,
-    selectedWhisperModel,
+    selectedWhisperDefId,
     needsKey,
     apiKeyForm,
     whisperStatus,
@@ -337,16 +330,14 @@ export default function OnboardingPage(): React.JSX.Element {
     }
   }, [llmCleanup, selectedLlm, needsLlmKey, llmKeyForm, navigate]);
 
-  const voiceModels = available.filter(
-    (m) => m.type === "voice" && CLOUD_VOICE_PROVIDERS.includes(m.provider_id),
-  );
-
-  const modelsByProvider = new Map<string, AvailableModel[]>();
-  for (const m of voiceModels) {
-    const list = modelsByProvider.get(m.provider_id) ?? [];
-    list.push(m);
-    modelsByProvider.set(m.provider_id, list);
-  }
+  const voiceItems = buildVoiceItems(available, whisperStatus, {
+    selectedModelId: selectedModel?.model_id,
+    selectedProvider:
+      selectedModel?.provider_id ??
+      (selectedWhisperDefId ? "local-whisper" : undefined),
+    selectedWhisperModelId: selectedWhisperDefId ?? undefined,
+    keyProviders: apiKeys,
+  });
 
   const llmModels = available.filter(
     (m) =>
@@ -363,7 +354,7 @@ export default function OnboardingPage(): React.JSX.Element {
   }
 
   const hasModelSelected =
-    selectedModel !== null || selectedWhisperModel !== null;
+    selectedModel !== null || selectedWhisperDefId !== null;
   const canAdvanceFromModel =
     hasModelSelected &&
     (!needsKey || apiKeyForm.watch("key").trim()) &&
@@ -544,290 +535,68 @@ export default function OnboardingPage(): React.JSX.Element {
                 </p>
               </div>
 
-              {/* Source toggle */}
-              <div className="flex justify-center">
-                <div className="border-border bg-secondary inline-flex rounded-md border p-[3px]">
-                  <button
-                    type="button"
-                    onClick={() => {
-                      setModelSource("cloud");
-                      setSelectedWhisperModel(null);
-                    }}
-                    className={cn(
-                      "flex items-center gap-1.5 rounded-[5px] px-3 py-1.5 text-[12px] transition-colors",
-                      modelSource === "cloud"
-                        ? "bg-card border-border text-foreground border font-medium shadow-sm"
-                        : "text-muted-foreground hover:text-foreground",
-                    )}
-                  >
-                    Cloud API
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => {
-                      setModelSource("local");
-                      setSelectedModel(null);
-                      setNeedsKey(false);
-                    }}
-                    className={cn(
-                      "flex items-center gap-1.5 rounded-[5px] px-3 py-1.5 text-[12px] transition-colors",
-                      modelSource === "local"
-                        ? "bg-card border-border text-foreground border font-medium shadow-sm"
-                        : "text-muted-foreground hover:text-foreground",
-                    )}
-                  >
-                    <HardDrive className="h-3 w-3" />
-                    Local
-                  </button>
-                </div>
+              {/* Unified voice model list */}
+              <div className="border-border max-h-[340px] overflow-y-auto rounded-[14px] border">
+                {voiceItems.length === 0 && (
+                  <div className="flex items-center gap-2 px-5 py-6">
+                    <Loader2 className="text-muted-foreground h-4 w-4 animate-spin" />
+                    <span className="text-muted-foreground text-sm">
+                      Loading models...
+                    </span>
+                  </div>
+                )}
+                {voiceItems.map((item, i) => (
+                  <VoiceRow
+                    key={item.key}
+                    item={item}
+                    first={i === 0}
+                    onSelectCloud={selectCloudModel}
+                    onSelectLocal={selectLocalModel}
+                    onDownload={downloadWhisperModel}
+                  />
+                ))}
               </div>
 
-              {modelSource === "cloud" && (
-                <>
-                  {/* Cloud model list */}
-                  <div className="border-border max-h-52 overflow-y-auto rounded-lg border">
-                    {[...modelsByProvider.entries()].map(
-                      ([providerId, models]) => (
-                        <div key={providerId}>
-                          <div className="text-muted-foreground bg-secondary/50 sticky top-0 px-3 py-1.5 text-[10px] font-semibold uppercase tracking-wider">
-                            {PROVIDER_DISPLAY_NAMES[providerId] ?? providerId}
-                          </div>
-                          {models.map((model) => (
-                            <button
-                              key={model.model_id}
-                              type="button"
-                              onClick={() => selectModel(model)}
-                              className={cn(
-                                "hover:bg-secondary flex w-full items-center gap-2 px-3 py-2 text-left text-sm",
-                                selectedModel?.model_id === model.model_id &&
-                                  "bg-primary/5",
-                              )}
-                            >
-                              <span className="flex-1">{model.model_name}</span>
-                              {selectedModel?.model_id === model.model_id && (
-                                <Check size={14} className="text-primary" />
-                              )}
-                            </button>
-                          ))}
-                        </div>
-                      ),
-                    )}
-                    {voiceModels.length === 0 && (
-                      <div className="flex items-center gap-2 px-3 py-4">
-                        <AlertTriangle className="text-muted-foreground h-4 w-4" />
-                        <span className="text-muted-foreground text-sm">
-                          Loading models...
-                        </span>
-                      </div>
-                    )}
-                  </div>
-
-                  {/* API key input */}
-                  {needsKey && selectedModel && (
-                    <div className="space-y-2">
-                      <p className="text-sm font-medium">
-                        Enter your{" "}
-                        {PROVIDER_DISPLAY_NAMES[selectedModel.provider_id] ??
-                          selectedModel.provider_id}{" "}
-                        API key
-                      </p>
-                      <div className="relative">
-                        <input
-                          type={showKey ? "text" : "password"}
-                          {...apiKeyForm.register("key")}
-                          placeholder="sk-..."
-                          className={cn(
-                            "border-border bg-card w-full rounded-lg border px-3 py-2.5 pr-10 font-mono text-sm",
-                            apiKeyForm.formState.errors.key &&
-                              "border-destructive",
-                          )}
-                          onKeyDown={(e) => {
-                            if (
-                              e.key === "Enter" &&
-                              apiKeyForm.getValues("key").trim()
-                            )
-                              saveVoiceAndContinue();
-                          }}
-                        />
-                        <button
-                          type="button"
-                          onClick={() => setShowKey(!showKey)}
-                          className="text-muted-foreground hover:text-foreground absolute right-3 top-1/2 -translate-y-1/2"
-                        >
-                          {showKey ? <EyeOff size={16} /> : <Eye size={16} />}
-                        </button>
-                      </div>
-                      {apiKeyForm.formState.errors.key && (
-                        <p className="text-destructive text-xs">
-                          {apiKeyForm.formState.errors.key.message}
-                        </p>
-                      )}
-                    </div>
-                  )}
-                </>
-              )}
-
-              {modelSource === "local" && (
-                <>
-                  <p className="text-muted-foreground text-xs leading-relaxed">
-                    Audio never leaves your device. Download a model to get
-                    started.
+              {/* API key input */}
+              {needsKey && selectedModel && (
+                <div className="space-y-2">
+                  <p className="text-sm font-medium">
+                    Enter your{" "}
+                    {PROVIDER_DISPLAY_NAMES[selectedModel.provider_id] ??
+                      selectedModel.provider_id}{" "}
+                    API key
                   </p>
-
-                  {whisperStatus?.binaryDownloading && (
-                    <div className="border-border bg-card flex items-center gap-2.5 rounded-lg border px-3 py-2.5">
-                      <Loader2 className="text-primary h-3.5 w-3.5 shrink-0 animate-spin" />
-                      <span className="text-muted-foreground text-xs">
-                        Building whisper.cpp from source...
-                      </span>
-                    </div>
-                  )}
-
-                  <div className="border-border max-h-52 overflow-y-auto rounded-lg border">
-                    {whisperStatus?.modelDefinitions.map((def) => {
-                      const state = whisperStatus.models.find(
-                        (m) => m.model === def.id,
-                      );
-                      const status = state?.status ?? "not_downloaded";
-                      const isSelected = selectedWhisperModel === def.id;
-
-                      return (
-                        <div
-                          key={def.id}
-                          className={cn(
-                            "border-border flex items-center gap-3 border-b px-3 py-2.5 last:border-b-0",
-                            isSelected && "bg-primary/5",
-                          )}
-                        >
-                          <div className="min-w-0 flex-1">
-                            <div className="flex items-center gap-2">
-                              <span className="text-foreground text-[13px] font-medium">
-                                {def.displayName}
-                              </span>
-                              <span className="text-muted-foreground text-[11px]">
-                                {formatBytes(def.sizeBytes)}
-                              </span>
-                              {def.quantized && (
-                                <span className="mono bg-primary/10 text-primary rounded-full px-1.5 py-[1px] text-[9px] tracking-wider">
-                                  FASTER
-                                </span>
-                              )}
-                            </div>
-                            <div className="text-muted-foreground mt-0.5 text-[11px]">
-                              {def.speed} · {def.quality} · {def.ramRequired}
-                            </div>
-
-                            {status === "downloading" &&
-                              state?.phase === "downloading_model" &&
-                              state.downloadProgress && (
-                                <div className="mt-1.5 space-y-1">
-                                  <div className="bg-secondary h-1.5 w-full overflow-hidden rounded-full">
-                                    <div
-                                      className="bg-primary h-full rounded-full transition-all"
-                                      style={{
-                                        width: `${state.downloadProgress.percent}%`,
-                                      }}
-                                    />
-                                  </div>
-                                  <div className="text-muted-foreground flex justify-between text-[10px]">
-                                    <span>
-                                      {formatBytes(
-                                        state.downloadProgress.bytesDownloaded,
-                                      )}{" "}
-                                      /{" "}
-                                      {formatBytes(
-                                        state.downloadProgress.bytesTotal,
-                                      )}
-                                    </span>
-                                    {state.downloadProgress.speedBps > 0 && (
-                                      <span>
-                                        {formatSpeed(
-                                          state.downloadProgress.speedBps,
-                                        )}
-                                      </span>
-                                    )}
-                                  </div>
-                                </div>
-                              )}
-
-                            {status === "downloading" &&
-                              state?.phase === "building_binary" && (
-                                <div className="mt-1.5">
-                                  <div className="bg-secondary h-1.5 w-full overflow-hidden rounded-full">
-                                    <div className="bg-primary h-full w-full animate-pulse rounded-full" />
-                                  </div>
-                                  <div className="text-muted-foreground mt-1 text-[10px]">
-                                    Building whisper.cpp...
-                                  </div>
-                                </div>
-                              )}
-
-                            {status === "error" && state?.error && (
-                              <div className="text-destructive mt-1 text-[11px]">
-                                {state.error}
-                              </div>
-                            )}
-                          </div>
-
-                          <div className="flex shrink-0 items-center gap-1.5">
-                            {(status === "not_downloaded" ||
-                              status === "error") && (
-                              <button
-                                type="button"
-                                onClick={() => downloadWhisperModel(def.id)}
-                                className="border-border hover:bg-secondary flex items-center gap-1.5 rounded-md border px-2.5 py-1.5 text-[11px] font-medium"
-                              >
-                                {status === "error" ? (
-                                  <>
-                                    <RefreshCw size={12} />
-                                    Retry
-                                  </>
-                                ) : (
-                                  <>
-                                    <Download size={12} />
-                                    Download
-                                  </>
-                                )}
-                              </button>
-                            )}
-                            {status === "downloading" && (
-                              <Loader2 className="text-muted-foreground h-4 w-4 animate-spin" />
-                            )}
-                            {status === "ready" && (
-                              <button
-                                type="button"
-                                onClick={() => selectLocalWhisper(def.id)}
-                                className={cn(
-                                  "rounded-md px-2.5 py-1.5 text-[11px] font-medium",
-                                  isSelected
-                                    ? "bg-primary text-primary-foreground"
-                                    : "border-border hover:bg-secondary border",
-                                )}
-                              >
-                                {isSelected ? (
-                                  <span className="flex items-center gap-1">
-                                    <Check size={12} /> Selected
-                                  </span>
-                                ) : (
-                                  "Use"
-                                )}
-                              </button>
-                            )}
-                          </div>
-                        </div>
-                      );
-                    })}
-
-                    {!whisperStatus && (
-                      <div className="flex items-center gap-2 px-3 py-4">
-                        <Loader2 className="text-muted-foreground h-4 w-4 animate-spin" />
-                        <span className="text-muted-foreground text-sm">
-                          Loading...
-                        </span>
-                      </div>
-                    )}
+                  <div className="relative">
+                    <input
+                      type={showKey ? "text" : "password"}
+                      {...apiKeyForm.register("key")}
+                      placeholder="sk-..."
+                      className={cn(
+                        "border-border bg-card w-full rounded-lg border px-3 py-2.5 pr-10 font-mono text-sm",
+                        apiKeyForm.formState.errors.key && "border-destructive",
+                      )}
+                      onKeyDown={(e) => {
+                        if (
+                          e.key === "Enter" &&
+                          apiKeyForm.getValues("key").trim()
+                        )
+                          saveVoiceAndContinue();
+                      }}
+                    />
+                    <button
+                      type="button"
+                      onClick={() => setShowKey(!showKey)}
+                      className="text-muted-foreground hover:text-foreground absolute right-3 top-1/2 -translate-y-1/2"
+                    >
+                      {showKey ? <EyeOff size={16} /> : <Eye size={16} />}
+                    </button>
                   </div>
-                </>
+                  {apiKeyForm.formState.errors.key && (
+                    <p className="text-destructive text-xs">
+                      {apiKeyForm.formState.errors.key.message}
+                    </p>
+                  )}
+                </div>
               )}
 
               <button
@@ -886,31 +655,7 @@ export default function OnboardingPage(): React.JSX.Element {
                       </p>
                     </div>
                   </div>
-                  <button
-                    type="button"
-                    onClick={() => setLlmCleanup(!llmCleanup)}
-                    className={cn(
-                      "relative h-[22px] w-10 shrink-0 rounded-full border transition-colors",
-                      llmCleanup
-                        ? "bg-primary border-primary/80"
-                        : "bg-secondary border-border",
-                    )}
-                    aria-pressed={llmCleanup}
-                  >
-                    <span
-                      className={cn(
-                        "absolute top-[1px] block h-[18px] w-[18px] rounded-full transition-transform",
-                        llmCleanup
-                          ? "bg-primary-foreground"
-                          : "bg-muted-foreground/70",
-                      )}
-                      style={{
-                        transform: llmCleanup
-                          ? "translateX(19px)"
-                          : "translateX(2px)",
-                      }}
-                    />
-                  </button>
+                  <Toggle on={llmCleanup} onChange={setLlmCleanup} />
                 </div>
               </div>
